+++ src/components/Logo/logoChangelogData.ts
@@ -10,6 +10,15 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Update",
+    affects: ["Styles"],
+    notes: [
+      'Udpates the "digitalCollectionsBlack" variant to use NYPL color design tokens.',
+    ],
+  },
+  {
     date: "2023-11-09",
     version: "2.1.2",
     type: "Update",